fix(core): resolve critical performance and safety issues identified in PR review
- C1: Optimized session history handling from O(n) deep clone to zero-copy reference borrow.
- C2: Cached system prompt at engine construction to eliminate redundant blocking I/O (git rev-parse) on the async runtime.
- C3: Standardized token counters to u64 and used saturating_add to prevent overflow on long sessions.

+++ src/api/anthropic.rs
@@ -126,12 +126,10 @@ impl LlmClient for AnthropicClient {
                             let usage = Usage {
                                 input_tokens: v["message"]["usage"]["input_tokens"]
                                     .as_u64()
-                                    .unwrap_or(0)
-                                    as u32,
+                                    .unwrap_or(0),
                                 output_tokens: v["message"]["usage"]["output_tokens"]
                                     .as_u64()
-                                    .unwrap_or(0)
-                                    as u32,
+                                    .unwrap_or(0),
                             };
                             Ok(StreamEvent::MessageStart {
                                 role,
@@ -193,10 +191,8 @@ impl LlmClient for AnthropicClient {
                                 v["delta"]["stop_reason"].as_str().map(|s| s.to_string());
                             let usage = if v["usage"].is_object() {
                                 Some(Usage {
-                                    input_tokens: v["usage"]["input_tokens"].as_u64().unwrap_or(0)
-                                        as u32,
-                                    output_tokens: v["usage"]["output_tokens"].as_u64().unwrap_or(0)
-                                        as u32,
+                                    input_tokens: v["usage"]["input_tokens"].as_u64().unwrap_or(0),
+                                    output_tokens: v["usage"]["output_tokens"].as_u64().unwrap_or(0),
                                 })
                             } else {
                                 None
